Add image dialog restore option, implement version icon missing logic and fix issues around remove links and rename dialog keyboard position

+++ app/src/main/java/mega/privacy/android/app/utils/Util.java
@@ -157,7 +157,7 @@ public class Util {
 	private static long lastClickTime;
 
 	// 150ms, a smaller value may cause the keyboard to fail to open
-	public final static int SHOW_IM_DELAY = 150;
+	public final static long SHOW_IM_DELAY = 150;
 
     /**
      * Language tag for simplified Chinese.
